Patch EF_2_acoustics.m to be octave-compatible (patch is not called the same way in Mlab and GNU/O)
Instead of just passing the c vector (colors based on vertices values) to patch as a color identifer,
use the mean value of c to determine the color of the face.
Note that it impacts slightly the rendering (there's no more interpolation of the color but only a flat color on the
whole face).